Make Cabin pieces visible
I've currently focusing on the 4 stone portions of the cabin. I can select them. they get replaced with a resource cube for stone. The rock is destroyed. The resource cube shrinks and is destroyed. On the screen, the name of a stone portion of the cabin appears. At this point, the stone portion is expected to be made visible.